requirements.txt
PyPDF2
uvicorn
fastapi
python-multipart
scikit-learn
python-docx
transformers
torch